requirements.txt
# requirements.txt
#autogen-agentchat @ git+https://github.com/patrickstolc/autogen.git@0.2#egg=autogen-agentchat
autogen==0.3.1
mistralai==1.2.3
ollama==0.3.3
fix-busted-json==0.0.18

# Added for web scraping and environment config
requests>=2.28.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
pip install google-search-results

